+++ lib/mailchimp_api/collection_parsers/tag.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module MailchimpAPI::CollectionParsers
+  class Tag < Base
+  end
+end